Audit fixes: Poly1305 tag, CRC32C table, packet validation, recv_buf sentinel, overflow UB
Bug fixes:
- Fix Poly1305 limb-to-word double-counting in poly1305_finish (RFC 8439 interop)
- Fix CRC32C software fallback table (entries 128-255 were corrupted)
- Add packet_type validation in nn_packet_write (reject > NN_PACKET_TYPE_MAX)
- Replace nn_recv_buf sentinel (NN_SEQ_EMPTY=0xFFFF) with occupied[] array
- Change nn_loss_window.index and nn_bandwidth.head from int to uint32_t (UB fix)

Quality:
- Reorder batch msg_len > UINT16_MAX check before capacity computation
- nn_rng_double: use 53 mantissa bits instead of 32
- Fix RNG comment (LCG + SplitMix output, not pure SplitMix)
- Remove dead ternary in nn_sent_buf_insert
- Pure FFI imports for seqGt/seqDiff (no IO wrapper)

Tests (335 → 520):
- Golden KAT for ChaCha20-Poly1305 AEAD
- CRC32C software fallback cross-check (NN_CRC32C_TEST_SW)
- Packet padding resilience + invalid type write
- Seq half-range boundary, 0xFFFF no-false-positive, ACK gap≥64,
  loss window wrap, sent buf collision, RNG seed 0, double extremes

+++ cbits/nn_bandwidth.c
@@ -18,7 +18,7 @@ nn_bandwidth_init(nn_bandwidth *bw, double window_ms)
 void
 nn_bandwidth_record(nn_bandwidth *bw, uint32_t size, uint64_t now_ns)
 {
-    int idx = bw->head & NN_BANDWIDTH_MASK;
+    uint32_t idx = bw->head & NN_BANDWIDTH_MASK;
     bw->timestamps[idx] = now_ns;
     bw->sizes[idx]      = size;
     bw->head++;
@@ -37,9 +37,9 @@ nn_bandwidth_bps(const nn_bandwidth *bw, uint64_t now_ns)
     uint64_t total_bytes = 0;
     uint64_t oldest = now_ns;
 
-    int start = bw->head - bw->count;
+    uint32_t start = bw->head - (uint32_t)bw->count;
     for (int i = 0; i < bw->count; i++) {
-        int idx = (start + i) & NN_BANDWIDTH_MASK;
+        uint32_t idx = (start + (uint32_t)i) & NN_BANDWIDTH_MASK;
         if (bw->timestamps[idx] >= cutoff) {
             total_bytes += bw->sizes[idx];
             if (bw->timestamps[idx] < oldest)
